minecat/pyproject.toml
[tool.poetry]
name = "minecat"
version = "0.1.0"
description = "A Discord bot to connect with Minecraft!"
authors = ["ooliver1 <oliverwilkes2006@icloud.com>"]
license = "ELv2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nextcord = {extras = ["speed"], version = "^2.2.0"}
ooliver-botbase = "^1.22.3"
docker-py = "^1.10.6"
redis = "^4.3.4"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
